renamed fsutil pkg to fileutil. moved WalkDir and IsNotEmpty funcs to fileutil pkg.
