+++ fileutil/fileutil.go
@@ -1,7 +1,11 @@
-package fsutil
+package fileutil
 
 import (
+	"errors"
+	"io/fs"
 	"os"
+	"path/filepath"
+	"syscall"
 )
 
 func IsExists(name string) (exists bool, err error) {
@@ -38,4 +42,31 @@ func IsDir(name string) (ok bool, err error) {
 func IsDir2(name string) (ok bool) {
 	ok, _ = IsDir(name)
 	return
+}
+
+func IsNotEmpty(err error) bool {
+	if e := new(os.PathError); errors.As(err, &e) {
+		if e.Err == syscall.ENOTEMPTY {
+			return true
+		}
+	}
+	return false
+}
+
+func WalkDir(root string, fn func(p string, d fs.DirEntry) bool) error {
+	return filepath.WalkDir(root, func(p string, d fs.DirEntry, e error) error {
+		if p == root {
+			return e
+		}
+		if e != nil {
+			if d != nil && d.IsDir() && os.IsNotExist(e) {
+				return fs.SkipDir
+			}
+			return e
+		}
+		if fn(p, d) {
+			return nil
+		}
+		return fs.SkipAll
+	})
 }
